refactor: remove old srpc_stub_gen.py and add new implementation for generating server and client stubs
